+++ app/src/main/java/dev/alllexey/itmowidgets/app/AppOverlayHostFragment.kt
@@ -63,6 +63,7 @@ internal val AppScreen.destinationId: Int
         AppScreen.DEBUG_TOOLS -> R.id.debug_tools
         AppScreen.RECORDBOOK_SUBJECT -> R.id.recordbook_subject
         AppScreen.APP_UPDATE -> R.id.app_update
+        AppScreen.QR_PASS -> R.id.qr_pass
         AppScreen.FRIENDS -> R.id.friends
         AppScreen.USER_FRIENDS -> R.id.user_friends
         AppScreen.USER_SEARCH -> R.id.user_search