+++ app/src/main/java/dev/alllexey/itmowidgets/feature/home/ui/HomeFragment.kt
@@ -7,6 +7,8 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import dagger.hilt.android.AndroidEntryPoint
 import dev.alllexey.itmowidgets.databinding.FragmentHomeBinding
+import dev.alllexey.itmowidgets.core.ui.navigation.AppScreen
+import dev.alllexey.itmowidgets.core.ui.navigation.openScreen
 
 @AndroidEntryPoint
 class HomeFragment : Fragment() {
@@ -28,6 +30,10 @@ class HomeFragment : Fragment() {
         _binding = FragmentHomeBinding.inflate(inflater, container, false)
         val view = binding.root
         return view
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        binding.qrFab.setOnClickListener { openScreen(AppScreen.QR_PASS) }
     }
 
     override fun onDestroyView() {