+++ app/src/main/java/dev/alllexey/itmowidgets/feature/qr/data/local/QrCodeLocalDataSource.kt
@@ -1,8 +1,11 @@
 package dev.alllexey.itmowidgets.feature.qr.data.local
 
 import kotlinx.coroutines.flow.Flow
+import dev.alllexey.itmowidgets.feature.qr.domain.QrCodeSnapshot
 
 interface QrCodeLocalDataSource {
+
+    fun snapshot(): QrCodeSnapshot?
 
     fun observe(): Flow<String>